refactor: Header Naming => WebRTCViewHeader
