--- src/Components/WebRTCView/Header/index.ts
@@ -1,1 +1,0 @@
-export { Header } from './Header';
