+++ src/Components/WebRTCView/WebRTCViewHeader/index.ts
@@ -1,0 +1,1 @@
+export { WebRTCViewHeader } from './WebRTCViewHeader';
